+++ package.json
@@ -29,9 +29,10 @@
   ],
   "scripts": {
     "build": "tsc --project tsconfig.json && tsc-alias -p tsconfig.json",
-    "clean": "node lib/clean.js",
     "fix": "eslint src/**/*.ts --fix",
+    "pr": "node lib/cli/pr.js",
     "pub": "npm publish --access public",
+    "pub-beta": "npm publish --tag beta",
     "sort": "npx sort-package-json"
   },
   "dependencies": {
